测试

+++ src/main/java/mining/tfidf/LSICal.java
@@ -1,7 +1,5 @@
 package mining.tfidf;
 
-import Jama.Matrix;
-import Jama.SingularValueDecomposition;
 import com.google.common.collect.BiMap;
 import lombok.extern.slf4j.Slf4j;
 import org.nd4j.linalg.api.ndarray.INDArray;
@@ -15,7 +13,7 @@ import java.util.Map;
 public class LSICal {
     private BiMap<Integer, Integer> termIdVocabularyId;
 
-    public double[][] transformMatrix(HashMap<Integer, HashMap<Integer, Double>> idTfIDf) {
+    public float[][] transformMatrix(HashMap<Integer, HashMap<Integer, Double>> idTfIDf) {
         BiMap<Integer, Integer> termIdVocabularyId = null;
         try {
             termIdVocabularyId = AllDocTfIdf.loadTermid();
@@ -24,14 +22,14 @@ public class LSICal {
         }
         BiMap<Integer, Integer> vocabularyIdTermId = termIdVocabularyId.inverse();
         log.info("正在生成" + idTfIDf.size() + "*" + termIdVocabularyId.size() + "的矩阵");
-        double[][] docTermMatrix = new double[idTfIDf.size()][termIdVocabularyId.size()];
+        float[][] docTermMatrix = new float[idTfIDf.size()][termIdVocabularyId.size()];
         for (Map.Entry<Integer, HashMap<Integer, Double>> entry : idTfIDf.entrySet()) {
             int docId = entry.getKey();
             for (Map.Entry<Integer, Double> termEntry : entry.getValue().entrySet()) {
                 int vocaId = termEntry.getKey();
                 double value = termEntry.getValue();
                 int termId = vocabularyIdTermId.get(vocaId);
-                docTermMatrix[docId][termId] = value;
+                docTermMatrix[docId][termId] = (float) value;
             }
         }
         return docTermMatrix;
@@ -43,7 +41,7 @@ public class LSICal {
      * @return 对于奇异值, 它跟我们特征分解中的特征值类似，在奇异值矩阵中也是按照从大到小排列，而且奇异值的减少特别的快，在很多情况下，前10%甚至1%的奇异值的和就占了全部的奇异值之和的99%以上的比例。也就是说，我们也可以用最大的k个的奇异值和对应的左右奇异向量来近似描述矩阵。也就是说：
      * Am×n=Um×mΣm×nVTn×n≈Um×kΣk×kVTk×n
      */
-    private double[][] svd(double[][] docTermMatrix) {
+    private float[][] svd(float[][] docTermMatrix) {
         log.info("开始进行SVD分解，得到Doc与Doc之间的关系");
 //        Matrix docTermM = new Matrix(docTermMatrix);
         //因为行的size要大于列的size，所以进行转置
